src/requirements.txt
pandas==1.4.2
python-dotenv==1.0.1
typer==0.12.3
scipy==1.8.0
numpy==1.22.3
matplotlib==3.5.2
matplotlib-inline==0.1.3
seaborn==0.11.2